vjepa2/requirements.txt
# torch>=2
# torchvision
tensorboard
wandb
iopath
pyyaml
numpy
opencv-python
submitit
braceexpand
webdataset
timm
transformers
peft
# decord
pandas
einops
beartype
psutil
h5py
fire
python-box
scikit-image
ftfy
jupyter
